Move code to lib/ as is more common, include ts code into npm module for reference, check in yarn.lock
