+++ package.json
@@ -7,10 +7,12 @@
   "scripts": {
     "build": "tsc",
     "lint": "tslint -p .",
-    "test": "mocha test/"
+    "test": "mocha test/",
+    "test-debug": "mocha test/ -b --no-exit"
   },
   "files": [
-    "dist/"
+    "dist",
+    "lib"
   ],
   "repository": {
     "type": "git",
